Roles updates and feature new project

+++ dotnet/businessservices/net.atos.daf.ct2.roleservicerest/Controllers/RoleController.cs
@@ -8,6 +8,8 @@ using RoleEntity=net.atos.daf.ct2.role.entity;
 using RoleComponent=net.atos.daf.ct2.role;
 using net.atos.daf.ct2.features;
 using net.atos.daf.ct2.features.entity;
+using net.atos.daf.ct2.roleservicerest.Entity;
+using System.Linq;
 
 namespace net.atos.daf.ct2.roleservicerest.Controllers
 {
@@ -27,22 +29,31 @@ namespace net.atos.daf.ct2.roleservicerest.Controllers
        
         [HttpPost]      
         [Route("create")]
-        public async Task<IActionResult> Create(RoleEntity.RoleMaster roleMaster)
+        public async Task<IActionResult> Create(Rolerequest roleMaster)
         {    
                try
                {
+
+                 if ((string.IsNullOrEmpty(roleMaster.RoleName)) || (roleMaster.OrganizationId == 0) )
+                {
+                    return StatusCode(400, "Role name and organization Id required.");
+                }
+                if(roleMaster.FeatureIds.Length ==  0 )
+                {
+                      return StatusCode(400, "Feature Ids required.");
+                }
                     RoleEntity.RoleMaster ObjRole = new RoleEntity.RoleMaster();
-                    ObjRole.Organization_Id =roleMaster.Organization_Id;
-                    ObjRole.Name = roleMaster.Name;
+                    ObjRole.Organization_Id =roleMaster.OrganizationId;
+                    ObjRole.Name = roleMaster.RoleName;
                     ObjRole.Createdby = roleMaster.Createdby;
                     ObjRole.Description = roleMaster.Description;
                     ObjRole.Feature_set_id=0;                  
                     
                     ObjRole.FeatureSet = new FeatureSet();
                     ObjRole.FeatureSet.Features = new List<Feature>();
-                    foreach(var item in roleMaster.FeatureSet.Features)
+                    foreach(var item in roleMaster.FeatureIds)
                     {
-                        ObjRole.FeatureSet.Features.Add(new Feature() { Id = item.Id });
+                        ObjRole.FeatureSet.Features.Add(new Feature() { Id = item });
                     }
                     int roleId = await roleManager.CreateRole(ObjRole);
                    return Ok(roleId);
@@ -56,15 +67,29 @@ namespace net.atos.daf.ct2.roleservicerest.Controllers
 
         [HttpPost]      
         [Route("update")]
-        public async Task<IActionResult> Update(RoleEntity.RoleMaster roleMaster)
+        public async Task<IActionResult> Update(Roleupdaterequest roleMaster)
         {    
+               if ((string.IsNullOrEmpty(roleMaster.RoleName)) || (roleMaster.OrganizationId == 0) ||(roleMaster.RoleId == 0)  )
+                {
+                    return StatusCode(400, "Role name and organization Id required Roleid required");
+                }
+                if(roleMaster.FeatureIds.Length ==  0 )
+                {
+                      return StatusCode(400, "Feature Ids required.");
+                }
                try
                {
                     RoleEntity.RoleMaster ObjRole = new RoleEntity.RoleMaster();                   
-                    ObjRole.Name = roleMaster.Name;
-                    ObjRole.Id = roleMaster.Id;
+                    ObjRole.Name = roleMaster.RoleName;
+                    ObjRole.Id = roleMaster.RoleId;
                     ObjRole.Description=roleMaster.Description;
-                    ObjRole.Updatedby = roleMaster.Updatedby;                    
+                    ObjRole.Updatedby = roleMaster.Updatedby;        
+                    ObjRole.FeatureSet = new FeatureSet();
+                    ObjRole.FeatureSet.Features = new List<Feature>();
+                    foreach(var item in roleMaster.FeatureIds)
+                    {
+                        ObjRole.FeatureSet.Features.Add(new Feature() { Id = item });
+                    }            
                     int roleId = await roleManager.UpdateRole(ObjRole);
                     logger.LogInformation(roleId+"Role Master Updated");
                     return Ok(roleId);
@@ -81,7 +106,12 @@ namespace net.atos.daf.ct2.roleservicerest.Controllers
         public async Task<IActionResult> Delete(int roleId, int updatedby)
         {    
                try
-               {                           
+               {         
+                if (roleId == 0) 
+                {
+                    return StatusCode(400, "Role id required ");
+                }
+                                    
                     int role_Id = await roleManager.DeleteRole(roleId,updatedby);
                     return Ok(role_Id);
                } 
@@ -98,11 +128,21 @@ namespace net.atos.daf.ct2.roleservicerest.Controllers
         {    
                try
                { 
+                 if (roleFilter.RoleId == 0 && roleFilter.Organization_Id == 0) 
+                {
+                    return StatusCode(400, "Role or organization Id required ");
+                }
                     var role = await roleManager.GetRoles(roleFilter); 
-                    List<RoleEntity.RoleMaster> roleList =new List<RoleEntity.RoleMaster>();
+                    List<Rolerequest> roleList =new List<Rolerequest>();
                     foreach(var roleitem in role)
                     {
-                         roleList.Add(roleitem);
+                         Rolerequest obj = new Rolerequest();
+                         obj.RoleId = roleitem.Id;
+                         obj.RoleName = roleitem.Name;
+                         obj.Description = roleitem.Description;
+                         obj.OrganizationId = roleitem.Organization_Id ?? 0;
+                         obj.FeatureIds = roleitem.FeatureSet.Features.Select(i=> i.Id).ToArray();
+                         roleList.Add(obj);
                     }
                                  
                   
@@ -111,7 +151,7 @@ namespace net.atos.daf.ct2.roleservicerest.Controllers
                catch(Exception ex)
                {
                     logger.LogError(ex.Message +" " +ex.StackTrace);
-                    return StatusCode(500,"Internal Server Error.");
+                    return StatusCode(500, ex.ToString());
                }
         }
